refactor: 추상화 불필요한 메서드 분리
